Implement separate packet and time queues
Putting data along with their new target timestamps in a priority queue
results in a lot of packet reordering that is more like packet
shuffeling.

This patch separates the packet and timestamp queue.
Generated timestamps are put in a priority queue so that timestamps are
guaranteed to be in order.
Packets are put in a regular FIFO queue so that packet order is
guaranteed to be unchanged.

The sender thread waits according to the topmost timestamp in the
timestamp queue as usual and then pops the topmost packet from the
packet queue.

This way it is guaranteed packets cannot be reordered or shuffled while
still allowing timestamp reordering and therefore very flexible
timestamp generation.
Since timestamps are still computed based on their original
receival-timestamp, each packet is still matched with an roughly
according timestamp.

This is probably the best trade-off we can get for modeling realistic
jitter.